requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
ftfy==6.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.24.0
idna==3.4
magic-filter==1.0.9
mailchecker==5.0.9
multidict==6.0.4
numpy==1.24.3
phonenumbers==8.13.11
psycopg2==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-fsutil==0.10.0
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
